Fixed the wp_query
My wp_query loop would query for all pages with the category
"front-page" in order to display specific pages. Unfortunately, the
query would search for the name of the page and append it to a
content-page where page would be replaced by the name of the posting. I
altered this code to search for the page's category and then set the
default categories to the section names that we desire. This way a user
can name a page anything they want for better readability and
personalization, all they have to do is set the category to whichever
template-part they desire.